Extracts authenticationUserController from userController
